// src/app/models/enums/ColorType.ts
export enum ColorType {
    PINK = 'PINK',
    PURPLE = 'PURPLE',
    RED = 'RED',
    BLUE = 'BLUE',
    GREEN = 'GREEN',
    YELLOW = 'YELLOW',
    ORANGE = 'ORANGE',
    TEAL = 'TEAL',
    INDIGO = 'INDIGO',
    GRAY = 'GRAY',
    BROWN = 'BROWN',
    CYAN = 'CYAN',
    LIME = 'LIME',
    AMBER = 'AMBER',
    EMERALD = 'EMERALD',
    FUCHSIA = 'FUCHSIA',
    MAGENTA = 'MAGENTA',
    MAROON = 'MAROON',
    OLIVE = 'OLIVE',
    NAVY = 'NAVY',
}